+++ plugin-flex-ts-template-v2/src/feature-library/contact-history/custom-components/ContactHistoryView/ContactHistoryView.Styles.ts
@@ -1,0 +1,10 @@
+import { styled } from '@twilio/flex-ui';
+
+export const ContactHistoryViewWrapper = styled('div')`
+  display: flex;
+  height: 100%;
+  overflow-y: scroll;
+  flex-flow: column;
+  flex-grow: 1;
+  flex-shrink: 1;
+`;
