+++ plugin-flex-ts-template-v2/src/feature-library/contact-history/flex-hooks/components/ViewCollection.tsx
@@ -1,7 +1,7 @@
 import * as Flex from '@twilio/flex-ui';
 
 import { FlexComponent } from '../../../../types/feature-loader';
-import ContactHistory from '../../custom-components/ContactHistoryView';
+import ContactHistory from '../../custom-components/ContactHistoryView/ContactHistoryView';
 
 export const componentName = FlexComponent.ViewCollection;
 export const componentHook = function addContactHistoryView(flex: typeof Flex) {
